Migrate to skylighting-core
To ensure that users of this library do not inadvertently pull in a
dependency on a GPL'd library (skylighting), this change modifies this
package to use and depend only on the new skylighting-core package, a
BSD-compatible subset of the skylighting functionality. The main thing
we give up in doing this is generated Haskell modules that provide
syntax definitions embedded in the program. Consequently users of this
library must load the syntax definitions themselves and provide them to
this library's functions. (skylighting could be used for that if
desired, but we no longer force users to do so.)